+++ tck/src/main/java/ee/jakarta/tck/data/standalone/nosql/example/NoSQLEntityTests.java
@@ -64,8 +64,8 @@ public class NoSQLEntityTests {
         //all nodes to get the latest update.
         TestPropertyUtility.waitForEventualConsistency(); 
         
-        int countExpensive = catalog.countByPriceGreaterThanEqual(2.99);
-        assertEquals(2, countExpensive, "Expected two products to be more than 3.00");
+        long countExpensive = catalog.countByPriceGreaterThanEqual(2.99);
+        assertEquals(2L, countExpensive, "Expected two products to be more than 3.00");
         
         Assertions.assertThrows(MappingException.class, () -> {
             catalog.countBySurgePriceGreaterThanEqual(2.99);
